[refs #223] Imports change due to moving the constants module
