+++ ozone/core/api/export_pdf/art7/section_imports.py
@@ -1,12 +1,12 @@
-from ..constants import TABLE_BLENDS_COMP_HEADER
-from ..constants import TABLE_BLENDS_COMP_STYLE
-from ..constants import TABLE_BLENDS_COMP_WIDTHS
-from ..constants import TABLE_IMPORTS_EXPORTS_HEADER
-from ..constants import TABLE_IMPORTS_EXPORTS_HEADER_STYLE
-from ..constants import TABLE_IMPORTS_EXPORTS_BL_WIDTHS
-from ..constants import TABLE_IMPORTS_EXPORTS_SUBS_WIDTHS
-from ..constants import TABLE_ROW_EMPTY_STYLE_IMP_EXP
-from ..constants import TABLE_ROW_EMPTY_IMP_EXP
+from .constants import TABLE_BLENDS_COMP_HEADER
+from .constants import TABLE_BLENDS_COMP_STYLE
+from .constants import TABLE_BLENDS_COMP_WIDTHS
+from .constants import TABLE_IMPORTS_EXPORTS_HEADER
+from .constants import TABLE_IMPORTS_EXPORTS_HEADER_STYLE
+from .constants import TABLE_IMPORTS_EXPORTS_BL_WIDTHS
+from .constants import TABLE_IMPORTS_EXPORTS_SUBS_WIDTHS
+from .constants import TABLE_ROW_EMPTY_STYLE_IMP_EXP
+from .constants import TABLE_ROW_EMPTY_IMP_EXP
 
 from ..util import get_comments_section
 from ..util import mk_table_blends
